+++ Samples~/Core/Scripts/Navigation/NavigationManager.cs
@@ -98,7 +98,7 @@ namespace Immersal.Samples.Navigation
 #if UNITY_EDITOR
                 if (instance == null && !Application.isPlaying)
                 {
-                    instance = FindObjectOfType<NavigationManager>();
+                    instance = FindFirstObjectByType<NavigationManager>();
                 }
 #endif
                 if (instance == null)
@@ -350,7 +350,7 @@ namespace Immersal.Samples.Navigation
         {
             if (m_XRSpace == null)
             {
-                m_XRSpace = FindObjectOfType<XRSpace>();
+                m_XRSpace = FindFirstObjectByType<XRSpace>();
 
                 if (m_XRSpace == null)
                 {